feat(hopr-lib): reorganize the API for easier edgli and gnosis_vpn-client usage (#7522)
