+++ hoprd/hoprd/Cargo.toml
@@ -18,19 +18,11 @@ default = [
   "transport-quic",
   "jemalloc-stats",
 ] # uses tokio by default because of axum
-prometheus = [
-  "dep:hopr-metrics",
-  "dep:lazy_static",
-  "hopr-lib/prometheus",
-  "hoprd-api/prometheus",
-  "hopr-network-types/prometheus",
-]
+prometheus = ["hopr-lib/prometheus", "hoprd-api/prometheus"]
 runtime-tokio = [
   "dep:tokio",
   "dep:tokio-retry",
   "hopr-lib/runtime-tokio",
-  "hopr-network-types/runtime-tokio",
-  "hopr-async-runtime/runtime-tokio",
   "opentelemetry_sdk?/rt-tokio",
 ]
 telemetry = [
@@ -55,9 +47,9 @@ anyhow = { workspace = true }
 async-lock = { workspace = true }
 async-signal = { workspace = true }
 clap = { workspace = true }
+console-subscriber = { workspace = true, optional = true }
 futures = { workspace = true }
 home = "0.5.11"
-lazy_static = { workspace = true, optional = true }
 proc-macro-regex = { workspace = true }
 serde = { workspace = true }
 serde_with = { workspace = true }
@@ -71,9 +63,6 @@ tokio-retry = { workspace = true, optional = true }
 tracing = { workspace = true }
 tracing-subscriber = { workspace = true, features = ["fmt", "json"] }
 validator = { workspace = true }
-
-console-subscriber = { workspace = true, optional = true }
-
 opentelemetry = { workspace = true, optional = true }
 opentelemetry-otlp = { workspace = true, default-features = false, features = [
   "trace",
@@ -89,18 +78,10 @@ hopr-lib = { workspace = true, features = [
   "runtime-tokio",
   "session-client",
   "session-server",
+  "utils_session",
+  "utils_parallel",
 ] }
 hoprd-api = { workspace = true }
-hopr-parallelize = { workspace = true }
-hopr-platform = { workspace = true }
-hopr-metrics = { workspace = true, optional = true }
-hopr-network-types = { workspace = true }
-hopr-async-runtime = { workspace = true }
-hopr-crypto-random = { workspace = true }
-
-[dev-dependencies]
-tempfile = { workspace = true }
-test-log = { workspace = true }
 
 # Target-specific dependencies for memory allocators
 [target.'cfg(target_os = "linux")'.dependencies]
@@ -110,3 +91,7 @@ tikv-jemalloc-ctl = { version = "0.6.0", features = ["stats"], optional = true }
 [build-dependencies]
 vergen-gix = { workspace = true, features = ["build"] }
 anyhow = { workspace = true }
+
+[dev-dependencies]
+tempfile = { workspace = true }
+test-log = { workspace = true }
