+++ hoprd/hoprd/src/main.rs
@@ -2,10 +2,13 @@ use std::{collections::HashMap, fmt::Formatter, str::FromStr, sync::Arc};
 
 use async_lock::RwLock;
 use async_signal::{Signal, Signals};
-use futures::{FutureExt, StreamExt, future::AbortHandle};
-use hopr_lib::{HoprKeys, HoprLibProcesses, IdentityRetrievalModes, ToHex};
+use futures::{
+    FutureExt, StreamExt,
+    future::{AbortHandle, abortable},
+};
+use hopr_lib::{HoprKeys, IdentityRetrievalModes, ToHex, state::HoprLibProcesses, utils::session::ListenerJoinHandles};
 use hoprd::{cli::CliArgs, errors::HoprdError, exit::HoprServerIpForwardingReactor};
-use hoprd_api::{ListenerJoinHandles, RestApiParameters, serve_api};
+use hoprd_api::{RestApiParameters, serve_api};
 use signal_hook::low_level;
 use tracing::{error, info, warn};
 use tracing_subscriber::prelude::*;
@@ -144,10 +147,6 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     #[cfg(all(target_os = "linux", feature = "jemalloc-stats"))]
     let _jemalloc_stats = jemalloc_stats::JemallocStats::start().await;
 
-    if hopr_crypto_random::is_rng_fixed() {
-        warn!("!! FOR TESTING ONLY !! THIS BUILD IS USING AN INSECURE FIXED RNG !!")
-    }
-
     if std::env::var("HOPR_TEST_DISABLE_CHECKS").is_ok_and(|v| v.to_lowercase() == "true") {
         warn!("!! FOR TESTING ONLY !! Node is running with some safety checks disabled!");
     }
@@ -236,7 +235,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
 
         processes.push(HoprdProcesses::ListenerSockets(session_listener_sockets.clone()));
 
-        processes.push(HoprdProcesses::RestApi(hopr_async_runtime::spawn_as_abortable!(
+        let (proc, abort_handle) = abortable(
             async move {
                 if let Err(e) = serve_api(RestApiParameters {
                     listener: api_listener,
@@ -251,8 +250,10 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
                     error!(error = %e, "the REST API server could not start")
                 }
             }
-            .inspect(|_| tracing::warn!(task = "hoprd - REST API", "long-running background task finished"))
-        )));
+            .inspect(|_| tracing::warn!(task = "hoprd - REST API", "long-running background task finished")),
+        );
+        let _jh = tokio::spawn(proc);
+        processes.push(HoprdProcesses::RestApi(abort_handle));
     }
 
     let (_hopr_socket, hopr_processes) = node
