+++ hoprd/rest-api/Cargo.toml
@@ -24,18 +24,15 @@ explicit-path = []
 runtime-tokio = []
 
 [dependencies]
-async-lock = { workspace = true }
 axum = { workspace = true, features = ["ws", "http2"] }
 axum-extra = { workspace = true, features = ["query"] }
 human-bandwidth = { workspace = true }
 base64 = { workspace = true }
 bytesize = { workspace = true }
 const_format = { workspace = true }
-dashmap = { workspace = true }
 either = { workspace = true }
 futures = { workspace = true }
 futures-concurrency = { workspace = true }
-futures-time = { workspace = true }
 lazy_static = { workspace = true, optional = true }
 regex = { workspace = true, optional = true }
 serde = { workspace = true }
@@ -54,17 +51,15 @@ utoipa-scalar = { workspace = true }
 utoipa-swagger-ui = { workspace = true }
 validator = { workspace = true }
 
-hopr-async-runtime = { workspace = true, features = ["runtime-tokio"] }
-hopr-lib = { workspace = true, features = ["runtime-tokio", "session-client"] }
-hopr-db-api = { workspace = true }
-hopr-crypto-types = { workspace = true }
-hopr-network-types = { workspace = true, features = ["runtime-tokio"] }
+hopr-lib = { workspace = true, features = [
+  "runtime-tokio",
+  "session-client",
+  "utils_session",
+  "utils_futures",
+] }
 hopr-metrics = { workspace = true, optional = true }
-hopr-platform = { workspace = true }
 
 [dev-dependencies]
 anyhow = { workspace = true }
 oas3 = { workspace = true }
 test-log = { workspace = true }
-
-hopr-transport-session = { workspace = true }
