+++ hoprd/rest-api/src/tickets.rs
@@ -5,10 +5,10 @@ use axum::{
     http::status::StatusCode,
     response::IntoResponse,
 };
-use hopr_crypto_types::types::Hash;
 use hopr_lib::{
     HoprBalance, HoprTransportError, ProtocolError, Ticket, TicketStatistics, ToHex,
     errors::{HoprLibError, HoprStatusError},
+    prelude::Hash,
 };
 use serde::Deserialize;
 use serde_with::{DisplayFromStr, serde_as};
