+++ transport/api/src/lib.rs
@@ -75,6 +75,9 @@ use hopr_transport_protocol::{
     errors::ProtocolError,
     processor::{MsgSender, PacketInteractionConfig, PacketSendFinalizer, SendMsgInput},
 };
+pub use hopr_transport_session as session;
+#[cfg(feature = "prometheus")]
+pub use hopr_transport_session::MetricSessionCounterGuard;
 #[cfg(feature = "runtime-tokio")]
 pub use hopr_transport_session::transfer_session;
 pub use hopr_transport_session::{
