+++ transport/session/src/lib.rs
@@ -19,6 +19,8 @@ pub use manager::{DispatchResult, MIN_SURB_BUFFER_DURATION, SessionManager, Sess
 pub use types::{
     ByteCapabilities, HoprSession, HoprSessionConfig, IncomingSession, ServiceId, SessionId, SessionTarget,
 };
+#[cfg(feature = "prometheus")]
+pub use utils::MetricSessionCounterGuard;
 #[cfg(feature = "runtime-tokio")]
 pub use utils::transfer_session;
